+++ examples/db_remember.rs
@@ -1,27 +1,23 @@
+// Set the `DB_REMEMBER_REDIS` environmental variable if you want to use Redis.
+// Otherwise, the default is Sqlite.
+
 use teloxide::{
-    dispatching2::dialogue::{serializer::Bincode, RedisStorage, Storage},
+    dispatching2::dialogue::{
+        serializer::{Bincode, Json},
+        ErasedStorage, RedisStorage, SqliteStorage, Storage,
+    },
     macros::DialogueState,
     prelude2::*,
     types::Me,
     utils::command::BotCommand,
-    RequestError,
 };
-use thiserror::Error;
 
-type MyDialogue = Dialogue<State, RedisStorage<Bincode>>;
-type StorageError = <RedisStorage<Bincode> as Storage<State>>::Error;
-
-#[derive(Debug, Error)]
-enum Error {
-    #[error("error from Telegram: {0}")]
-    TelegramError(#[from] RequestError),
-
-    #[error("error from storage: {0}")]
-    StorageError(#[from] StorageError),
-}
+type MyDialogue = Dialogue<State, ErasedStorage<State>>;
+type MyStorage = std::sync::Arc<ErasedStorage<State>>;
+type HandlerResult = Result<(), Box<dyn std::error::Error + Send + Sync>>;
 
 #[derive(DialogueState, Clone, serde::Serialize, serde::Deserialize)]
-#[handler_out(anyhow::Result<()>)]
+#[handler_out(HandlerResult)]
 pub enum State {
     #[handler(handle_start)]
     Start,
@@ -44,17 +40,22 @@ pub enum Command {
     #[command(description = "reset your number.")]
     Reset,
 }
+
 #[tokio::main]
 async fn main() {
+    pretty_env_logger::init();
+    log::info!("Starting db_remember_bot...");
+
     let bot = Bot::from_env().auto_send();
-    // You can also choose serializer::JSON or serializer::CBOR
-    // All serializers but JSON require enabling feature
-    // "serializer-<name>", e. g. "serializer-cbor"
-    // or "serializer-bincode"
-    let storage = RedisStorage::open("redis://127.0.0.1:6379", Bincode).await.unwrap();
+
+    let storage: MyStorage = if std::env::var("DB_REMEMBER_REDIS").is_ok() {
+        RedisStorage::open("redis://127.0.0.1:6379", Bincode).await.unwrap().erase()
+    } else {
+        SqliteStorage::open("db.sqlite", Json).await.unwrap().erase()
+    };
 
     let handler = Update::filter_message()
-        .enter_dialogue::<Message, RedisStorage<Bincode>, State>()
+        .enter_dialogue::<Message, ErasedStorage<State>, State>()
         .dispatch_by::<State>();
 
     Dispatcher::builder(bot, handler)
@@ -65,11 +66,7 @@ async fn main() {
         .await;
 }
 
-async fn handle_start(
-    bot: AutoSend<Bot>,
-    msg: Message,
-    dialogue: MyDialogue,
-) -> anyhow::Result<()> {
+async fn handle_start(bot: AutoSend<Bot>, msg: Message, dialogue: MyDialogue) -> HandlerResult {
     match msg.text().unwrap().parse() {
         Ok(number) => {
             dialogue.update(State::GotNumber(number)).await?;
@@ -93,7 +90,7 @@ async fn handle_got_number(
     dialogue: MyDialogue,
     num: i32,
     me: Me,
-) -> anyhow::Result<()> {
+) -> HandlerResult {
     let ans = msg.text().unwrap();
     let bot_name = me.user.username.unwrap();